View, edit and delete posts in the X demo
Editing re-opens the SAME editor with the post's content, and the id
option is what routes the save back to the right row — the button reads
Save rather than Post, and nothing about the editor changes.

Delete confirms inside the action sheet rather than through a system
alert. Chaining an alert off the sheet's dismissal never appeared at all:
iOS silently refuses to present one while a sheet is on its way out. A
destructive confirmation inside the sheet is what iOS apps do anyway, and
it has no presentation conflict to lose to.

Also fixes the relative time, which read 0.478572s — Carbon returns a
float from diffInSeconds now.

+++ app/Models/Post.php
@@ -36,7 +36,7 @@ class Post extends Model
     /** "2h", "3d" — the compact relative time a timeline uses. */
     public function age(): string
     {
-        $seconds = max(0, $this->created_at?->diffInSeconds() ?? 0);
+        $seconds = (int) max(0, $this->created_at?->diffInSeconds() ?? 0);
 
         return match (true) {
             $seconds < 60 => $seconds.'s',
